refactor(utils): remove util/utils
