+++ app/components/Wallet/wallethelper.ts
@@ -1,0 +1,3 @@
+export function trimAddress(address: string): string {
+	return `${address.substring(0, 6)}...${address.substring(address.length - 4)}`;
+}
